+++ src/_sample/app/layout.tsx
@@ -1,15 +1,15 @@
-import sampleProvider from '../lib/providers/sampleProvider';
+// import sampleProvider from '../lib/providers/sampleProvider';
 
-export default function Layout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body>
-        <sampleProvider>{children}</sampleProvider>
-      </body>
-    </html>
-  );
-}
+// export default function Layout({
+//   children,
+// }: Readonly<{
+//   children: React.ReactNode;
+// }>) {
+//   return (
+//     <html lang="en">
+//       <body>
+//         <sampleProvider>{children}</sampleProvider>
+//       </body>
+//     </html>
+//   );
+// }
