+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,32 +1,5 @@
+import { authOptions } from '@/lib/utils/auth';
 import NextAuth from 'next-auth';
-import { PrismaAdapter } from '@auth/prisma-adapter';
-import GoogleProvider from 'next-auth/providers/google';
-import type { NextAuthOptions } from 'next-auth';
-import { prisma } from '@/lib/prisma';
-import { PATH } from '@/constants/path';
-
-// NextAuth 설정 옵션을 정의합니다.
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID || '',
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET || '',
-    }),
-  ],
-  callbacks: {
-    async session({ session, user }) {
-      if (session.user) {
-        session.user.id = user.id;
-      }
-      return session;
-    },
-  },
-
-  pages: {
-    signIn: PATH.HOME,
-  },
-};
 
 const handler = NextAuth(authOptions);
